pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
authors = [{name = "Lucas Faudman", email = "lucasfaudman@gmail.com"}]
dynamic = ["version"]
name = "apk_secret_scanner"
requires-python = ">=3.8"
readme = "README.md"
license = { file = "LICENSE.txt" }
description = "Decompile and scan for secrets in .apk, .dex, .jar, .class, .smali, .zip, .aar, .arsc, .aab, .xapk, .jadx.kts files"
keywords = ["secret", "scanner", "jadx", "decompile", "apk", "dex", "jar", "class", "smali", "zip", "aar", "arsc", "aab", "xapk", "jadx.kts"]
dependencies = [
    "pyyaml"
]

[project.urls]
Homepage = "https://github.com/LucasFaudman/apk-secret-scanner.git"
Repository = "https://github.com/LucasFaudman/apk-secret-scanner.git"

[tool.setuptools]

[tool.setuptools_scm]
version_file = "src/apk_secret_scanner/_version.py"

[tool.black]
line-length = 120
